+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # HEAD
 
+- Added a Sprint BTS profile selectable with `--bts-profile sprint`.
 - Mobile IPv4 replies now provide DNS settings.
 - Added authenticated Mobile IPv4 registration for carrier profiles.
 - Improved Band Class 1 (i.e. Sprint).